requirements.txt
sentence-transformers==2.7.0
faiss-cpu==1.8.0
numpy==1.24.3
tiktoken==0.7.0
torch>=2.0.0
transformers>=4.32.0
google-generativeai>=0.3.0
# Optional for reranking evaluation
matplotlib>=3.5.0
seaborn>=0.11.0
pandas>=1.3.0
